pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "image-review"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "matplotlib",
    "numpy",
    "pydicom",
    "scikit-image",
    "rectpack",
    "tqdm",
    "pygame-ce",
]

[project.scripts]
image-review = "image_review.cli:main"
